#8 - fix: rename Message comp
