+++ src/pages/LikePage/LikePage.tsx
@@ -1,10 +1,10 @@
-import { Message } from '../../components/Message/Message';
-import MessageStructure from '../../types/Message';
+import { ProfileMessage } from '../../components/ProfileMessage/ProfileMessage';
+import Message from '../../types/Message';
 import './LikePage.css';
 import { FaRegHeart } from 'react-icons/fa';
 
 interface LikePageProps {
-  messagesData: MessageStructure[];
+  messagesData: Message[];
   handleClickLike: (messageId: number) => void;
   handleClickBookmark: (messageId: number) => void;
   handleClickDelete: (messageId: number) => void;
@@ -28,7 +28,7 @@ export const LikePage = ({
     <>
       {filterMessagesData.length > 0 ? (
         sortedMessages.map((message) => (
-          <Message
+          <ProfileMessage
             key={message.id}
             message={message}
             onClickLike={handleClickLike}
